+++ Back-end/src/main/java/com/simplon/dvdStore/dto/SaleDTO.java
@@ -4,9 +4,9 @@ package com.simplon.dvdStore.dto;
  * The type Sale dto.
  */
 public record SaleDTO(
-        long client,
-        long dvd,
-        int quantity
+        Integer client,
+        Integer dvd,
+        Integer quantity
 ) {
 }
